Implemented (part of) hierarchy for Expressions
Implemented the basis of the hierarchy for expressions (See #116): added
superclass for expressions and statements (Executable) and made
ArgumentExpression -> ArgumentExecutable because Statements ==
Expressions without return type, therefore we can reuse code for both of
them in ArgumentExecutable.

Still needs lots of tests (but wanted to see if this hierarchy is OK
first).